requirements.txt
kafka-python
feast
tensorflow
bentoml
pandas
numpy
